==> 5_vader_sentiment_analysis.ipynb <==
import pandas as pd

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer



fb2019 = pd.read_json('fb.json')

#preprocess texts (remove urls, mentions and e-mail addresses, as well as html markup)
text = fb2019["post"].astype(str) 

#remove html markup
text =text.str.replace('<.*?>|&([a-z0-9]+|#[0-9]{1,6}|#x[0-9a-f]{1,6});','')

#remove tabs
text= text.str.replace('[ |\t]{2,}', '', case=False)

#remove urls 
text= text.str.replace('http\S+|www.\S+', '', case=False)

#remove additional links
text= text.str.replace('bit\.ly\S+', '', case=False)

text = text.tolist()

#perform sentiment analysis

analyzer = SentimentIntensityAnalyzer()

vs_res=[]
for row in text:
    vs = analyzer.polarity_scores(row)
    vs_res.append(vs)

list_vader = [item for item in vs_res]
vader_score =[item['compound'] for item in list_vader]
fb =fb.assign(vader_score = pd.Series(vader_score))

fb.to_json('fb.json')
