==> 6_stance_classification.ipynb <==
import pandas as pd
import numpy as np
import tensorflow as tf
from tensorflow import keras
import pickle

from sentence_transformers import SentenceTransformer
from transformers import AutoTokenizer

from joblib import load

import glob
import os

import keras

from numpy import genfromtxt


fb = pd.read_json ("fb.json")
fb

#clean strings

text = fb['post']

#remove html markup
text = text.str.replace('<.*?>|&([a-z0-9]+|#[0-9]{1,6}|#x[0-9a-f]{1,6});',' ')

#remove tabs
text= text.str.replace('[ |\\t]{2,}', ' ', case=False, regex = True)

#remove urls 
text= text.str.replace('http\\S+|www\\.\\S+', ' ', case=False, regex = True)

#remove urls 
text= text.str.replace('\\S+\\.se|\\S+\\.com|\\S+\\.org', ' ', case=False, regex = True)

#remove e-mails
text= text.str.replace('\\S+@\\S+', ' ', case=False, regex = True)

#remove additional links
text= text.str.replace('bit\\.ly\\S+', ' ', case=False, regex = True)
text= text.str.replace('goo\\.gl\\S+', ' ', case=False, regex = True)
text= text.str.replace('youtu\\.be\\S+', ' ', case=False, regex = True)

#remove characters after semicolon
text= text.str.replace('\\:\\S+', '', case=False, regex = True)

#replace ... with .
text = text.str.replace('\\.\\.+','\\.', regex = True)

text = text.str.replace(' +',' ').str.strip()

text = text.tolist()

model = SentenceTransformer('KBLab/sentence-bert-swedish-cased')

column_ids = []
for i in range(0, 768):
    column_ids.append(str(i))

#generate features for the unseen data using BERT
import csv

embeddings = []
for i in range(0, len (text)):
    
    
    encoding = model.encode(text[i])
    embeddings.append(encoding)
    
    if (i>0) & (i % 200000 == 0):
        print(i)
        
        df = pd.DataFrame(embeddings, columns = column_ids)
        df.to_csv("embeddings" + str(i) + ".csv")

        embeddings = []
    
    elif (i == (len(text)-1)):
        
        df = pd.DataFrame(embeddings, columns = column_ids)
        df.to_csv("embeddings" + str(i) + ".csv")

#load previously fine-tuned model to classify on- and off-topic messages
tf_model = keras.models.load_model('model_bert_topic_offtopic')

#classify messages

all_files = ['embeddings200000.csv', 'embeddings400000.csv', 'embeddings600000.csv', 'embeddings800000.csv', 
             'embeddings1000000.csv', 'embeddings1200000.csv', 'embeddings1400000.csv', 'embeddings1600000.csv',  
            'embeddings1800000.csv', 'embeddings2000000.csv', 'embeddings2002635.csv']

class_names = ['off-topic', 'topic']
labels = []
values = []

for filename in all_files:
    
    print(filename)
    embeddings = pd.read_csv(filename) 
    embeddings = embeddings.drop (columns = "Unnamed: 0").to_numpy()
    embeddings_X = np.expand_dims(embeddings, axis=1)
    predictions = tf_model(embeddings_X, training=False)
    print(len(predictions))
   
    for i, logits in enumerate(predictions):
       
        class_idx=np.argmax(logits.numpy())
        name = class_names[class_idx]
        val = np.max(logits.numpy())
    
        labels.append(name)
        values.append (val)


#CLASSIFY STANCEs WITH THE FINE-TUNED BERT MODEL¶

#load model to classify stances (negative and non-negative)
clf = load("trained_classifier_model_mlp-sbert-features_negative_non-negative.joblib")

#classify messages
all_files = ['embeddings200000.csv', 'embeddings400000.csv', 'embeddings600000.csv', 'embeddings800000.csv', 
             'embeddings1000000.csv', 'embeddings1200000.csv', 'embeddings1400000.csv', 'embeddings1600000.csv',  
            'embeddings1800000.csv', 'embeddings2000000.csv', 'embeddings2002635.csv']

stance_names = ['neg', 'nonneg']
stance_probs = []
stance_labels = []

for filename in all_files:
    
    print(filename)
    embeddings = pd.read_csv(filename) 
    embeddings = embeddings.drop (columns = "Unnamed: 0").to_numpy()
    
    labels_pred = clf.predict(embeddings)
    probs_pred = clf.predict_proba(embeddings) 
   
    for i, logits in enumerate(probs_pred):
     
        class_idx=np.argmax(logits)
        name = stance_names[class_idx]    
        val = np.max(logits)
    
        stance_labels.append(name)
        stance_probs.append (val)


df = pd.DataFrame(
    {'tot_label': labels,
     'tot_prob': values,
     'stance_label': stance_labels,
     'stance_prob' : stance_probs
    })

df.to_csv("flashback_stance_classification.csv")

#add data on user stances to the df

fb = pd.merge(fb, df, left_index = True, right_index=True)

fb['stance_label'] = np.where(fb['tot_label'] == "off-topic", fb['tot_label'], fb['stance_label'])
fb['stance_prob'] = np.where(fb['stance_label'] == "off-topic", fb['tot_prob'], fb['stance_prob'])
fb.to_json("fb.json")
