==> 9_conversation_properties.ipynb <==
import pandas as pd
from itertools import chain
from collections import Counter

from urllib.parse import urlparse
import re
import numpy as np
import requests
import matplotlib.pyplot as plt
from matplotlib import pyplot
from matplotlib.pyplot import figure
import math

import seaborn as sns

from scipy.stats import shapiro 
from scipy.stats import lognorm
from scipy.stats import mannwhitneyu

import scipy.stats as stats

import pymannkendall
from statsmodels.tsa.stattools import grangercausalitytests

from statsmodels.tsa.stattools import adfuller
from statsmodels.tsa.stattools import kpss

import statsmodels.api as sm

import statsmodels.formula.api as smf

#Load the data
##############################################


fb=pd.read_json("fb.json")

#upload the network and edgelist data depending on the time interval chosen (14 / 27 / 54 min)
g = pd.read_csv("conversations_graph_27min.csv", names = ["convid", "n_nodes", "n_edges", "ideg", "odeg", "deg", "avg_clust", "recip", "dens"])
g.convid = g.convid.astype("str")

e = pd.read_csv ("conversations_edges_27min.csv", names = ["convid", "sender", "recepient", "timestamp"])
e

el = pd.read_csv ("conversation_edgelist27min.csv")


#Fix conversation ids and dates
################################################

fb['delta'] = (fb['date']-fb['date'].shift()).fillna(pd.Timedelta('0 days'))
fb['delta'] = fb['delta'].dt.total_seconds() / 60
fb.delta = fb.delta.astype(int)

fb["convid"] = fb["convid"].astype(str)
fb.convid = fb.convid.apply(lambda x: re.sub("\\.0", "", x))
convids = fb.convid.unique().tolist()
convids.remove("nan")

fb = fb.assign (date24h = fb["date"].dt.floor("24h")).reset_index (drop = True)

e.timestamp = pd.to_datetime(e.timestamp, format="%Y-%m-%d %H:%M")
e.timestamp

el.time_sender = pd.to_datetime(el.time_sender, format="%Y-%m-%d %H:%M")
el.time_recepient = pd.to_datetime(el.time_recepient, format="%Y-%m-%d %H:%M")
el  = el.merge(e[["convid", "sender", "recepient"]], left_on = ["sender", "recepient"], right_on = ["sender", "recepient"])

# SUMMARY STATS

#Create CONVERSATION statistics (summary df)
###############################################
fb_conv = fb[fb.convid != "nan"].groupby("convid").agg({"num_am": 'sum', "num_mm": 'sum', "num_all": 'sum', 
                                                            "vader_score": 'mean', 'stance_label':
                                                            'count'}).rename (columns ={"stance_label": "num_posts"})

#links
fb_conv['am_density'] = fb_conv.num_am.div(fb_conv ["num_posts"])
fb_conv['mm_density'] = fb_conv.num_mm.div(fb_conv["num_posts"])
fb_conv['all_density'] = fb_conv.num_all.div(fb_conv ["num_posts"])

fb_conv['am_to_all'] = fb_conv.num_am.div(fb_conv ["num_all"])
fb_conv['mm_to_all'] = fb_conv.num_mm.div(fb_conv ["num_all"])

#stances

threads_neg = fb[(fb.convid != "nan")&(fb.stance_label == "neg")].groupby ("convid").agg({'stance_label' : 'count',
                                                                                              'stance_prob' : 'mean'})
threads_neg = threads_neg.rename (columns={"stance_label": "stance_label_neg", "stance_prob": "mean_stance_prob_neg"})


#POSITIVE STANCES

threads_nonneg = fb[(fb.convid != "nan")&(fb.stance_label == "nonneg")].groupby ("convid").agg({'stance_label' : 
                                                                                                    'count', 
                                                                                                    'stance_prob' : 
                                                                                                    'mean'})
threads_nonneg = threads_nonneg.rename (columns={"stance_label": "stance_label_nonneg", "stance_prob": 
                                                 "mean_stance_prob_nonneg"})

#
fb_conv = pd.concat([fb_conv, threads_neg, threads_nonneg], axis = 1)

fb_conv['neg_label_ratio'] = fb_conv.stance_label_neg.div(fb_conv ["num_posts"])
fb_conv['nonneg_label_ratio'] = fb_conv.stance_label_nonneg.div(fb_conv ["num_posts"])

fb_conv = fb_conv.assign (convid = fb_conv.index).reset_index (drop = True)


#conversation durations
diffdf = (fb[fb.convid != "nan"].groupby(['convid'])['date']
         .agg(lambda x: x.iat[-1] - x.iat[0])
         .reset_index(name='diff'))

el_sumdiff = el[["convid", "time_diff"]].groupby ("convid", as_index=False).median("time_diff")
el_sumdiff.convid = el_sumdiff.convid.astype(str)
el_sumdiff.convid = el_sumdiff.convid.apply(lambda x: re.sub("\\.0", "", x))


#fb_conv = pd.concat([fb_conv, diffdf], axis = 1)
fb_conv = fb_conv.merge(diffdf, right_on = "convid", left_on = "convid")


fb_conv["diff_hours"] = fb_conv["diff"]/ np.timedelta64(1, 'h')

fb_conv = fb_conv.merge(el_sumdiff, left_on = "convid", right_on = "convid")

#add network metrics

fb_conv = fb_conv.merge(g, left_on = "convid", right_on = "convid")

print(fb_conv)

print(np.mean(fb_conv.num_posts))
print(np.median(fb_conv.num_posts))


fb_conv.to_excel("fb_conv_27.xlsx")

# Mann-Whitney U test - LINKS

# AM VERSUS MM CONVERSATIONS
######################################################

am_conv = fb_conv[(fb_conv.am_density>0.2) & (fb_conv.mm_density == 0)] #other values tested: 0.2, 0.4, 0.6
print(len(am_conv))
mm_conv = fb_conv[(fb_conv.mm_density>0.2) & (fb_conv.am_density == 0)] #other values tested: 0.2, 0.4, 0.6
print(len(mm_conv))

#conversations with am links are longer?
mw_hours = mannwhitneyu(x=am_conv.diff_hours, y=mm_conv.diff_hours, alternative='greater') #‘greater’: the distribution underlying x is stochastically greater than the distribution underlying y, i.e. F(u) < G(u) for all u.
print(mw_hours)

#conversations with am links have higher sentiments? +++
mw_vader = mannwhitneyu(x=am_conv.vader_score, y=mm_conv.vader_score, alternative='greater') 
print(mw_vader)

#conversations with am links have higher negative stance probability? 
mw_negprob = mannwhitneyu(x=am_conv.mean_stance_prob_neg.dropna(), y=mm_conv.mean_stance_prob_neg.dropna(), alternative='greater') 
print(mw_negprob)

#conversations with am links have lower non-negative stance probability? 
mw_nonnegprob = mannwhitneyu(x=am_conv.mean_stance_prob_nonneg.dropna(), y=mm_conv.mean_stance_prob_nonneg.dropna(), alternative='less') 
print(mw_nonnegprob)

#conversations with am links have higher negative stance ratio? ++++
mw_negprob = mannwhitneyu(x=am_conv.neg_label_ratio.dropna(), y=mm_conv.neg_label_ratio.dropna(), alternative='greater') 
print(mw_negprob)

#conversations with am links have lower non-negative stance ratio?
mw_nonnegprob = mannwhitneyu(x=am_conv.nonneg_label_ratio.dropna(), y=mm_conv.nonneg_label_ratio.dropna(), alternative='less') 
print(mw_nonnegprob)

#conversations with am links have lower clustering?
mw_clust = mannwhitneyu(x=am_conv.avg_clust, y=mm_conv.avg_clust, alternative='less') 
print(mw_clust)

#conversations with am links have lower num nodes?
mw_nodes = mannwhitneyu(x=am_conv.n_nodes, y=mm_conv.n_nodes, alternative='less') 
print(mw_nodes)

#conversations with am links have lower num edges?
mw_edges = mannwhitneyu(x=am_conv.n_edges, y=mm_conv.n_edges, alternative='less') 
print(mw_edges)

#conversations with am links have higher density?
mw_dens = mannwhitneyu(x=am_conv.dens, y=mm_conv.dens, alternative='greater') 
print(mw_dens)


# CONVERSATIONS WITHOUT AND WITH ANY TYPE OF LINKS
#######################################################

all_conv = fb_conv[fb_conv.all_density>0.6] #other values tested: 0.2, 0.4, 0.6

none_conv = fb_conv[fb_conv.all_density==0]
#none_conv = none_conv[0:300]    

print(len(all_conv))
print(len(none_conv))

#conversations with any links are longer? ++++++
mw_hours = mannwhitneyu(x=all_conv.diff_hours, y=none_conv.diff_hours, alternative='greater') #‘greater’: the distribution underlying x is stochastically greater than the distribution underlying y, i.e. F(u) < G(u) for all u.
print(mw_hours)

#conversations with any links have lower sentiments? 
mw_vader = mannwhitneyu(x=all_conv.vader_score, y=none_conv.vader_score, alternative='less') 
print(mw_vader)

#conversations with any links have higher negative stance probability?  +++++
mw_negprob = mannwhitneyu(x=all_conv.mean_stance_prob_neg.dropna(), y=none_conv.mean_stance_prob_neg.dropna(), alternative='less') 
print(mw_negprob)

#conversations with any links have lower non-negative stance probability? +++++ 
mw_nonnegprob = mannwhitneyu(x=all_conv.mean_stance_prob_nonneg.dropna(), y=none_conv.mean_stance_prob_nonneg.dropna(), alternative='greater') 
print(mw_nonnegprob)

#conversations with any links have lower negative negative stance ratio? +++++
mw_negprob = mannwhitneyu(x=all_conv.neg_label_ratio.dropna(), y=none_conv.neg_label_ratio.dropna(), alternative='less') 
print(mw_negprob)

#conversations with any links have higher non-negative stance ratio? +++++
mw_nonnegprob = mannwhitneyu(x=all_conv.nonneg_label_ratio.dropna(), y=none_conv.nonneg_label_ratio.dropna(), alternative='greater') 
print(mw_nonnegprob)

#conversations with any links have higher clustering? 
mw_clust = mannwhitneyu(x=all_conv.avg_clust, y=none_conv.avg_clust, alternative='greater') 
print(mw_clust)

#conversations with any links have lhigher num nodes? +++++++
mw_nodes = mannwhitneyu(x=all_conv.n_nodes, y=none_conv.n_nodes, alternative='greater') 
print(mw_nodes)

#conversations with any links have higher num edges? +++++++++
mw_edges = mannwhitneyu(x=all_conv.n_edges, y=none_conv.n_edges, alternative='greater') 
print(mw_edges) 

#conversations with am links have higher density? ++++++
mw_dens = mannwhitneyu(x=all_conv.dens, y=none_conv.dens, alternative='less') 
print(mw_dens)


# ALT RIGHT LINKS

#how many conversations have any link?
print(len(fb.convid[fb.num_all>0].unique())/len(fb.convid.unique()))

#how many conversations have an am link?
print(len(fb.convid[fb.num_am>0].unique())/len(fb.convid.unique()))

#how many conversations have a mm link?
print(len(fb.convid[fb.num_mm>0].unique())/len(fb.convid.unique()))


#how many unique users share am content?

print(len(fb.userid.unique()))
print(len(fb.userid[fb.num_am>0].unique()))
print(len(fb.userid[fb.num_am>0].unique())/len(fb.userid.unique()))
print(len(fb.userid.unique()))
